1007tool/requirements.txt
pygame==2.6.1
gitpython==3.1.41
mechanize==0.4.8
